ts to js w/ authController working
