--- src/server/models/userSchema.ts
@@ -1,33 +1,0 @@
-export {};
-const mongoose = require('mongoose');
-
-const userSchema = mongoose.Schema({
-  username: { 
-    type: String, 
-    required: true 
-  },
-  email: { 
-    type: String, 
-    required: true, 
-    unique: true
-  },
-  password: { 
-    type: String, 
-    required: true 
-  },
-  karma: { 
-    type: Number, 
-    required: true
-  },
-  tasks: [
-    {
-      type: mongoose.Types.ObjectId, 
-      require: true, 
-      ref: 'Task'
-    }
-  ]
-});
-
-// const User = mongoose.model('User', userSchema);
-
-module.exports = mongoose.model('User', userSchema);;
